Refactor testgrid file locations
