+++ BUILD
@@ -56,8 +56,7 @@ filegroup(
         "//robots/commenter:all-srcs",
         "//robots/issue-creator:all-srcs",
         "//scenarios:all-srcs",
-        "//testgrid/config:all-srcs",
-        "//testgrid/updater:all-srcs",
+        "//testgrid:all-srcs",
         "//triage:all-srcs",
         "//velodrome:all-srcs",
         "//vendor:all-srcs",
