+++ testgrid/cmd/config/BUILD
@@ -7,52 +7,28 @@
 package(default_visibility = ["//visibility:public"])
 
 load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library", "go_test")
-load("@io_bazel_rules_go//proto:def.bzl", "go_proto_library")
-
-proto_library(
-    name = "config_proto",
-    srcs = ["config.proto"],
-)
-
-go_proto_library(
-    name = "config_pb",
-    importpath = "k8s.io/test-infra/testgrid/config/config_pb",
-    proto = ":config_proto",
-)
-
-genrule(
-    name = "gen-config",
-    srcs = [
-        "config.yaml",
-    ],
-    outs = ["testgrid-config"],
-    cmd = "$(location //testgrid/config) --yaml=$< --output=$@",
-    tools = [
-        "//testgrid/config",
-    ],
-)
 
 go_binary(
     name = "config",
     embed = [":go_default_library"],
-    importpath = "k8s.io/test-infra/testgrid/config",
+    importpath = "k8s.io/test-infra/testgrid/cmd/config",
 )
 
 go_test(
     name = "go_default_test",
     srcs = ["config_test.go"],
     data = [
-        "config.yaml",
         "//jenkins/job-configs:all-srcs",
         "//mungegithub:submit-queue/deployment/kubernetes/configmap.yaml",
         "//prow:config.yaml",
+        "//testgrid:config-yaml",
     ],
     embed = [":go_default_library"],
-    importpath = "k8s.io/test-infra/testgrid/config",
+    importpath = "k8s.io/test-infra/testgrid/cmd/config",
     deps = [
-        ":config_pb",
         "//prow/config:go_default_library",
-        "//testgrid/config/yaml2proto:go_default_library",
+        "//testgrid:config",
+        "//testgrid/yaml2proto:go_default_library",
         "//vendor/github.com/ghodss/yaml:go_default_library",
     ],
 )
@@ -60,8 +36,8 @@ go_test(
 go_library(
     name = "go_default_library",
     srcs = ["main.go"],
-    importpath = "k8s.io/test-infra/testgrid/config",
-    deps = ["//testgrid/config/yaml2proto:go_default_library"],
+    importpath = "k8s.io/test-infra/testgrid/cmd/config",
+    deps = ["//testgrid/yaml2proto:go_default_library"],
 )
 
 filegroup(
@@ -73,9 +49,6 @@ filegroup(
 
 filegroup(
     name = "all-srcs",
-    srcs = [
-        ":package-srcs",
-        "//testgrid/config/yaml2proto:all-srcs",
-    ],
+    srcs = [":package-srcs"],
     tags = ["automanaged"],
 )
